Refactor: owner 경로 나누고, principal으로 검증

+++ backend/src/main/java/com/ssafy/keeping/domain/menu/service/MenuService.java
@@ -10,6 +10,8 @@ import com.ssafy.keeping.domain.menuCategory.repository.MenuCategoryRepository;
 import com.ssafy.keeping.domain.store.model.Store;
 import com.ssafy.keeping.domain.store.repository.StoreRepository;
 import com.ssafy.keeping.domain.store.service.StoreService;
+import com.ssafy.keeping.domain.user.owner.model.Owner;
+import com.ssafy.keeping.domain.user.owner.repository.OwnerRepository;
 import com.ssafy.keeping.global.exception.CustomException;
 import com.ssafy.keeping.global.exception.constants.ErrorCode;
 import org.springframework.transaction.annotation.Transactional;
@@ -26,17 +28,18 @@ import java.util.Optional;
 public class MenuService {
     private final MenuRepository menuRepository;
     private final StoreRepository storeRepository;
+    private final OwnerRepository ownerRepository;
     private final MenuCategoryRepository menuCategoryRepository;
 
-    public MenuResponseDto createMenu(Long storeId, MenuRequestDto requestDto) {
-        Store store = storeRepository.findById(storeId).orElseThrow(
-                () -> new CustomException(ErrorCode.STORE_NOT_FOUND)
-        );
+    public MenuResponseDto createMenu(Long ownerId, Long storeId, MenuRequestDto requestDto) {
+        Owner owner = validOwner(ownerId);
+        Store store = validStore(storeId);
+
+        if (!store.getOwner().getOwnerId().equals(owner.getOwnerId()))
+            throw new CustomException(ErrorCode.OWNER_NOT_MATCH);
 
         Long categoryId = requestDto.getCategoryId();
-        MenuCategory category = menuCategoryRepository.findById(categoryId).orElseThrow(
-                () -> new CustomException(ErrorCode.MENU_CATEGORY_NOT_FOUND)
-        );
+        MenuCategory category = validMenuCategory(categoryId);
 
         if(!Objects.equals(storeId, category.getStore().getStoreId())){
             throw new CustomException(ErrorCode.STORE_NOT_MATCH);
@@ -64,12 +67,16 @@ public class MenuService {
         );
     }
 
-    public MenuResponseDto editMenu(Long storeId, Long menuId, MenuEditRequestDto requestDto) {
-        Menu menu = menuRepository.findByMenuIdAndStore_StoreId(menuId, storeId)
-                .orElseThrow(() -> new CustomException(ErrorCode.MENU_NOT_FOUND));
+    public MenuResponseDto editMenu(Long ownerId, Long storeId, Long menuId, MenuEditRequestDto requestDto) {
+        Owner owner = validOwner(ownerId);
+        Store store = validStore(storeId);
 
-        MenuCategory category = menuCategoryRepository.findById(requestDto.getCategoryId())
-                .orElseThrow(() -> new CustomException(ErrorCode.MENU_CATEGORY_NOT_FOUND));
+        if (!store.getOwner().getOwnerId().equals(owner.getOwnerId()))
+            throw new CustomException(ErrorCode.OWNER_NOT_MATCH);
+
+        Menu menu = validMenu(menuId, storeId);
+
+        MenuCategory category = validMenuCategory(requestDto.getCategoryId());
         if (!category.getStore().getStoreId().equals(storeId))
             throw new CustomException(ErrorCode.STORE_NOT_MATCH);
 
@@ -104,18 +111,48 @@ public class MenuService {
         return menuRepository.findAllMenusByStoreId(storeId);
     }
 
-    public void deleteMenu(Long storeId, Long menusId) {
-        Menu menu = menuRepository.findByMenuIdAndStore_StoreId(menusId, storeId)
-                .orElseThrow(() -> new CustomException(ErrorCode.MENU_NOT_FOUND));
+    public void deleteMenu(Long ownerId, Long storeId, Long menusId) {
+        Owner owner = validOwner(ownerId);
+        Store store = validStore(storeId);
+
+        if (!store.getOwner().getOwnerId().equals(owner.getOwnerId()))
+            throw new CustomException(ErrorCode.OWNER_NOT_MATCH);
+
+        Menu menu = validMenu(menusId, storeId);
 
         menuRepository.deleteById(menusId);
     }
 
-    public void deleteAllMenu(Long storeId) {
-        storeRepository.findById(storeId).orElseThrow(
-                () -> new CustomException(ErrorCode.STORE_NOT_FOUND)
-        );
+    public void deleteAllMenu(Long ownerId, Long storeId) {
+        Owner owner = validOwner(ownerId);
+        Store store = validStore(storeId);
+
+        if (!store.getOwner().getOwnerId().equals(owner.getOwnerId()))
+            throw new CustomException(ErrorCode.OWNER_NOT_MATCH);
 
         menuRepository.deleteAllByStore_StoreId(storeId);
+    }
+
+    private Owner validOwner(Long ownerId) {
+        return ownerRepository.findById(ownerId).orElseThrow(
+                () -> new CustomException(ErrorCode.OWNER_NOT_FOUND)
+        );
+    }
+
+    private Store validStore(Long storeId) {
+        return storeRepository.findById(storeId).orElseThrow(
+                () -> new CustomException(ErrorCode.STORE_NOT_FOUND)
+        );
+    }
+
+    private MenuCategory validMenuCategory(Long categoryId) {
+        return  menuCategoryRepository.findById(categoryId).orElseThrow(
+                () -> new CustomException(ErrorCode.MENU_CATEGORY_NOT_FOUND)
+        );
+    }
+
+    private Menu validMenu(Long menuId, Long storeId) {
+        return menuRepository.findByMenuIdAndStore_StoreId(menuId, storeId)
+                .orElseThrow(() -> new CustomException(ErrorCode.MENU_NOT_FOUND));
     }
 }
